Add a table of contents.
The table of contents is dynamically generated from the actual content section using the `jekyll-toc` snippet. (This snippet is available under an MIT license so it can be directly incorporated into this site under the same terms.)

This also includes styling for a sticky sidebar that contains the TOC. It's hidden on mobile.